+++ src/sll/sandbox.c
@@ -8,6 +8,7 @@
 #include <sll/object.h>
 #include <sll/sandbox.h>
 #include <sll/scheduler.h>
+#include <sll/static_object.h>
 #include <sll/thread.h>
 #include <sll/types.h>
 
@@ -16,8 +17,7 @@
 #define FLAG(f,nm) __STATIC_STRING(__flag_str_##f,(nm))
 #define CHECK_FLAG(f) \
 	if (fl&(1ull<<SLL_SANDBOX_FLAG_##f)){ \
-		o->v[i]=sll_create_object(SLL_OBJECT_TYPE_STRING); \
-		sll_string_clone(&__flag_str_##f,&(o->v[i]->dt.s)); \
+		o->v[i]=sll_string_to_object(&__flag_str_##f); \
 		i++; \
 	}
 #define CHECK_SET_FLAG(f) \